+++ lib/play.h
@@ -1,11 +1,11 @@
 #ifndef PLAY_H
 #define PLAY_H
 
-//#include "main.h"
+#include "main.h"
 #include "maps.h"
 
 /* Maximum player inventory */
-#define MAX_INV 32
+#define MAX_INV 40
 
 /* Main player structure */
 struct player {
@@ -19,6 +19,6 @@ struct player {
 extern void	player_init(struct player *cur_player);
 extern void	player_quit(struct player *cur_player);
 extern void	move_player(struct worldmap *map, struct player *cur_player, int x, int y);
-
+extern void	toggle_inv(struct game *cur_game);
 
 #endif